refactor: infra 모듈 생성. firebase 클래스 이동
