+++ infra/src/main/java/com/kbe5/infra/config/FcmConfig.java
@@ -1,4 +1,5 @@
-package com.kbe5.common.config;
+package com.kbe5.infra.config;
+
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
