+++ infra/src/main/java/com/kbe5/infra/firebase/dto/UpdateFcmTokenRequest.java
@@ -1,4 +1,4 @@
-package com.kbe5.domain.commonservice.firebase.dto;
+package com.kbe5.infra.firebase.dto;
 
 public record UpdateFcmTokenRequest(
         String token
