+++ rento-adapter/build.gradle
@@ -16,6 +16,7 @@ configurations {
 dependencies {
     implementation project(':common')
     implementation project(':domain')
+    implementation project(':infra')
 
     // Swagger OpenAPI 3
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.7.0'
