+++ rento-adapter/src/main/java/com/kbe5/adapter/AdapterApplication.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @SpringBootApplication(scanBasePackages = {
         "com.kbe5.common",
         "com.kbe5.domain",
+        "com.kbe5.infra",
         "com.kbe5.adapter"
 }
 )
