+++ rento-adapter/src/main/java/com/kbe5/adapter/service/DriveService.java
@@ -1,4 +1,4 @@
-package com.kbe5.domain.commonservice;
+package com.kbe5.adapter.service;
