+++ rento-api/build.gradle
@@ -18,6 +18,8 @@ dependencies {
     //modules
     implementation project(':common')
     implementation project(':domain')
+    implementation project(':infra')
+
 
     // Swagger OpenAPI 3
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.7.0'
