+++ rento-api/src/main/java/com/kbe5/api/ApiApplication.java
@@ -9,6 +9,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @SpringBootApplication(scanBasePackages = {
         "com.kbe5.common",
         "com.kbe5.domain",
+        "com.kbe5.infra",
         "com.kbe5.api"
 })
 @EnableJpaRepositories(basePackages = "com.kbe5.domain")
