+++ settings.gradle
@@ -3,3 +3,4 @@ include 'rento-api'
 include 'rento-adapter'
 include 'common'
 include 'domain'
+include 'infra'
